# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SmartHub AgroChain Platform API
  description: Enterprise REST API specification for SmartHub AgroChain — Agricultural Marketplace, Financial Settlement, and Escrow Platform.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Environment
  - url: https://api.smarthub-agrochain.com/api
    description: Production Environment
paths:
...
  /payments/webhook:
    post:
      summary: Idempotent Payment Settlement Webhook
      description: Handles payment gateway callbacks, executes fee split, releases escrow, and emits PAYMENT_COMPLETED event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transactionRef:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Payment processed idempotently.
        '400':
          description: Invalid signature or missing transactionRef.